refactor: extract profile components
